Create RelevantEvents<> struct to store relevant events This commit is a major refactor of the storage and passing around of events across functions.
The main change is the separation of concerns of EventStorage.
Now, EventStorage is a trait for a simple key-value store of epoch IDs to events.
RelevantEvents handles filtering the stored events by relevancy, and holding them in a struct that can be easily passed around.
Previously, EventStorage handled both storage and relevancy filtering.

Doing so enables the following other improvements:
- clear semantics: "no relevant events" is now always an empty array. No more checking for Ok(Some(None)).
- Removed EpochEvents trait, as it is no longer needed. This also reduces the FFI surface that needs to be implemented.
- Removed a bunch of generics and phantom fields from PrivateDataService.
- compute_epoch_source_losses() now takes a set of source URIs, not the entire mapping of URIs to events.