Implement batching algorithm and release filter (#89)
* Implement batching algorithm and release filter
Created new ReleaseFilter (supertrait of Filter) for filters that can release budget over time.
Created PureDPBudgetReleaseFilter, which is the first implementor of ReleaseFilter.

Other small changes:
- PrivateDataService no longer uses/depends on anyhow
- PpaHistogramRequest::new() takes &config as a reference, to reduce cloning

* Pre-initialize non-global filters in batch_pds
Ideally we would store 2 types of filters in a single FilterStorage: release filters for C, and regular PurePDBudgetFilters for the others.
As batch_pds code is still experimental, this code works for now.

Other changes:
- add comments to reflect which rust code reflects which pseudocode line in the paper

* Create init_default_logging() function for unit tests

---------

Co-authored-by: Pierre Tholoniat <pierre@tholoniat.com>