Allow EventStorage to implement its own relevant event selection Adds a new EventStorage::relevant_events_for_epoch(), with a default implementation that matches the previous behaviour. This means that the code is still backwards-compatible, and behaves exactly the same. However, the EventStorage can now override the default implementation if it has a more optimised implementation.
The main use-case for this is HashMapEventStorage. Before, when requesting the relevant events for epoch 6, it would clone ALL the events for that epoch, and discard the irrelevant ones.
Now, it can filter them beforehand and only clone the relevant ones.

There is an argument to be made that EventStorage is a thin, simple trait and wouldn't benefit from extra functionality. However in my opinion, its sibling trait FilterStorage benefits a lot from being slightly more than an FFI wrapper, so there is precedent.